lesson 3: Xử lý đưa dự liệu board vào trong redux, fix bug, refactor code
