+++ src/apis/index.js
@@ -2,11 +2,6 @@ import axios from 'axios'
 import { API_ROOT } from '~/utils/constants'
 
 /* APIs Board */
-export const fetchBoardDetailsAPI = async (boardId) => {
-  const response = await axios.get(`${API_ROOT}/v1/boards/${boardId}`)
-  return response.data
-}
-
 export const updateBoardDetailsAPI = async (boardId, boardData) => {
   const response = await axios.put(`${API_ROOT}/v1/boards/${boardId}`, boardData)
   return response.data
