+++ src/pages/Boards/_id.jsx
@@ -1,122 +1,65 @@
-import { useEffect, useState } from 'react'
+import { useEffect } from 'react'
+
+import { useSelector, useDispatch } from 'react-redux'
+import { fetchBoardDetailsAPI, updateCurrentActiveBoard, selectCurrentActiveBoard } from '~/redux/activeBoard/activeBoardSlice'
+
 import Container from '@mui/material/Container'
 import AppBar from '~/components/AppBar/AppBar'
 import BoardBar from './BoardBar/BoardBar'
 import BoardContent from './BoardContent/BoardContent'
-import { generatePlaceholderCard } from '~/utils/formatters'
-import { isEmpty } from 'lodash'
-import { mapOrder } from '~/utils/sorts'
 import Box from '@mui/material/Box'
 import CircularProgress from '@mui/material/CircularProgress'
 import Typography from '@mui/material/Typography'
-import { toast } from 'react-toastify'
+import { cloneDeep } from 'lodash'
 
-// import { mockData } from '~/apis/mock-data'
 import {
-  fetchBoardDetailsAPI,
-  createNewColumnAPI,
-  createNewCardAPI,
   updateBoardDetailsAPI,
   updateColumnDetailsAPI,
-  moveCardToDifferentColumnAPI,
-  deleteColumnDetailsAPI
+  moveCardToDifferentColumnAPI
 } from '~/apis/index'
 
 
+
 const Board = () => {
-  const [board, setBoard] = useState(null)
+  const dispatch = useDispatch()
+  const board = useSelector(selectCurrentActiveBoard)
 
   useEffect(() => {
     const boardId = '65a215706b8948401c354667'
-    fetchBoardDetailsAPI(boardId).then(board => {
-      board.columns = mapOrder(board?.columns, board?.columnOrderIds, '_id')
-
-      board.columns.forEach(column => {
-        if (isEmpty(column.cards)) {
-          column.cards = [generatePlaceholderCard(column)]
-          column.cardOrderIds = [generatePlaceholderCard(column)._id]
-        } else {
-          column.cards = mapOrder(column?.cards, column?.cardOrderIds, '_id')
-        }
-      })
-
-      setBoard(board)
-    })
-  }, [])
-
-  /* API Columns */
-  const createNewColumn = async (newColumnData) => {
-    const newColumn = await createNewColumnAPI({
-      ...newColumnData,
-      boardId: board._id
-    })
-
-    newColumn.cards = [generatePlaceholderCard(newColumn)]
-    newColumn.cardOrderIds = [generatePlaceholderCard(newColumn)._id]
-
-    const newBoard = { ...board }
-    newBoard.columns.push(newColumn)
-    newBoard.columnOrderIds.push(newColumn._id)
-    setBoard(newBoard)
-  }
-
-  /* API Cards */
-  const createNewCard = async (newCardData) => {
-    const newCard = await createNewCardAPI({
-      ...newCardData,
-      boardId: board._id
-    })
-
-    const newBoard = { ...board }
-    const updateColumn = newBoard.columns.find(column => column._id === newCard.columnId)
-
-    if (updateColumn) {
-      if (updateColumn.cards.some(card => card.FE_placeholder_card)) {
-        //  remove placeholder card when column have 1 card
-        updateColumn.cards = [newCard]
-        updateColumn.cardOrderIds = [newCard._id]
-      } else {
-        updateColumn.cards.push(newCard)
-        updateColumn.cardOrderIds.push(newCard._id)
-      }
-    }
-
-    setBoard(newBoard)
-  }
+    dispatch(fetchBoardDetailsAPI(boardId))
+  }, [dispatch])
 
   const moveColumn = (dndOrderedColumns) => {
     const dndOrderedColumnsIds = dndOrderedColumns.map(c => c._id)
 
-    const newBoard = { ...board }
+    const newBoard = cloneDeep(board)
     newBoard.columns = dndOrderedColumns
     newBoard.columnOrderIds = dndOrderedColumnsIds
-    setBoard(newBoard)
+    dispatch(updateCurrentActiveBoard(newBoard))
 
     updateBoardDetailsAPI(board._id, { columnOrderIds: dndOrderedColumnsIds })
   }
 
   const moveCard = (dndOrderedCards, dndOrderedCardIds, columnId) => {
-    const newBoard = { ...board }
+    const newBoard = cloneDeep(board)
     const columnUpdate = newBoard.columns.find(column => column._id === columnId)
     if (columnUpdate) {
       columnUpdate.cards = dndOrderedCards
       columnUpdate.cardOrderIds = dndOrderedCardIds
-      setBoard(newBoard)
+      dispatch(updateCurrentActiveBoard(newBoard))
     }
-
     updateColumnDetailsAPI(columnId, { cardOrderIds: dndOrderedCardIds })
   }
 
   const moveCardToDifferentColumn = (currentCardId, prevColumnId, nextColumnId, dndOrderedColumns) => {
     const dndOrderedColumnsIds = dndOrderedColumns.map(c => c._id)
 
-    const newBoard = { ...board }
+    const newBoard = cloneDeep(board)
     newBoard.columns = dndOrderedColumns
     newBoard.columnOrderIds = dndOrderedColumnsIds
-    setBoard(newBoard)
+    dispatch(updateCurrentActiveBoard(newBoard))
 
     let prevCardOrderIds = dndOrderedColumns.find(column => column._id === prevColumnId)?.cardOrderIds
-    // remove placeholder card when column is empty
     if (prevCardOrderIds[0].includes('placeholder-card')) prevCardOrderIds = []
 
     moveCardToDifferentColumnAPI({
@@ -126,15 +69,6 @@ const Board = () => {
       nextColumnId,
       nextCardOrderIds: dndOrderedColumns.find(column => column._id === nextColumnId)?.cardOrderIds
     })
-  }
-
-  const deleteColumn = (columnId) => {
-    const newBoard = { ...board }
-    newBoard.columns = newBoard.columns.filter(column => column._id !== columnId)
-    newBoard.columnOrderIds = newBoard.columnOrderIds.filter(_id => _id !== columnId)
-    setBoard(newBoard)
-
-    deleteColumnDetailsAPI(columnId).then((result) => toast.success(result.deleteResult))
   }
 
   if (!board) {
@@ -159,12 +93,9 @@ const Board = () => {
       <BoardBar board={board} />
       <BoardContent
         board={board}
-        createNewColumn={createNewColumn}
-        createNewCard={createNewCard}
         moveColumn={moveColumn}
         moveCard={moveCard}
         moveCardToDifferentColumn={moveCardToDifferentColumn}
-        deleteColumn={deleteColumn}
       />
     </Container>
   )
